[REFACTOR] Move HTML-STF and IntelliJ-STF parts to STF-Project
I moved the html STF classes from ui project to the new stf project. The
Saros Eclipse project (where the test are located) now depends on the
stf project. In addition I moved the IntelliJHTMLWorkbenchBot to the STF
project.
