+++ de.fu_berlin.inf.dpp.stf/.classpath
@@ -3,11 +3,9 @@
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.6"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src/html"/>
-	<classpathentry kind="src" path="src/eclipse"/>
 	<classpathentry kind="src" path="src/intellij"/>
-	<classpathentry exported="true" kind="lib" path="lib/swt-browser-improved.jar" sourcepath="lib/swt-browser-improved-src.zip"/>
-	<classpathentry kind="lib" path="test/lib/easymock-3.1.jar"/>
-	<classpathentry kind="lib" path="test/lib/powermock-easymock-1.5.1-full.jar"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.junit.JUNIT_CONTAINER/4"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/de.fu_berlin.inf.dpp.intellij"/>
+	<classpathentry kind="lib" path="/de.fu_berlin.inf.dpp.intellij/lib/fest-swing-1.2.1.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
