+++ de.fu_berlin.inf.dpp.stf/src/html/de/fu_berlin/inf/dpp/stf/server/rmi/htmlbot/widget/IRemoteHTMLTextElement.java
