+++ de.fu_berlin.inf.dpp.ui/.classpath
@@ -4,7 +4,6 @@
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry excluding="**/package-info.java" kind="src" path="src"/>
 	<classpathentry kind="src" path="test/junit"/>
-	<classpathentry kind="src" path="test/framework/stf/src"/>
 	<classpathentry kind="lib" path="lib/gson-2.3.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/swt-browser-improved.jar" sourcepath="lib/swt-browser-improved-src.zip"/>
 	<classpathentry kind="lib" path="test/lib/easymock-3.1.jar"/>
