+++ de.fu_berlin.inf.dpp.ui/build.properties
@@ -1,5 +1,4 @@
-source.. = src/,\
-           test/framework/stf/src/
+source.. = src/,
 output.. = bin/
 bin.includes = META-INF/,\
                .,\
